further refactoring

+++ python/perf-kernels/tools/plot-layout/plot_layout.py
@@ -1,62 +1,82 @@
 import argparse
-import sys
 import os
+import sys
 
 from blocked import generate_blocked_tex
 from dot import generate_dot_tex
 from lds import generate_lds_tex
+from utils import run_bash_command, OneLineFormatter
 from wmma import generate_wmma_tex
-from utils import run_bash_command
 
 
 def parse_args():
-    parser = argparse.ArgumentParser(
+    top_parser = argparse.ArgumentParser(
         prog="Draw triton layouts",
         allow_abbrev=True,
     )
-    parser.add_argument("--output", type=str, default="myplot", help='output pdf file name (without surfix)')
-    parser.add_argument("--keep", action='store_true', default=False, help='If set, keep the generated .tex file')
-    subparsers = parser.add_subparsers(
+    top_parser.add_argument("--output", type=str, default="myplot", help='output pdf file name (without surfix)')
+    top_parser.add_argument("--keep", action='store_true', default=False, help='If set, keep the generated .tex file')
+    top_parser.add_argument("--force", action='store_true', default=False, help='If set, overwrite the pdf file with the same name')
+    subparsers = top_parser.add_subparsers(
         dest="plot_type",
         metavar="PLOT_TYPE",
-        required=True,
+        required=False,
         title="subcommands",
         description="Choose to plot blocked, lds, dot or wmma",
         help="Choose one of the four plot mode"
     )
-    ## blocked layout parameters
-    blocked_parser = subparsers.add_parser("blocked", help="plot blocked layout for global memory access")
-        ## tensor shapes
-    blocked_parser.add_argument("--tensorShape", type=int, nargs=2, default=(128, 64),
-                        help='2D block shape in the form of (dim0, dim1)')
-    blocked_parser.add_argument("--rowName", type=str, default="M", help='tensor dim0 name')
-    blocked_parser.add_argument("--colName", type=str, default="K", help='tensor dim1 name')
-    blocked_parser.add_argument("--sizePerThread", type=int, nargs=2, default=(1, 4))
-    blocked_parser.add_argument("--threadsPerWarp", type=int, nargs=2, default=(16, 4))
-    blocked_parser.add_argument("--warpsPerCTA", type=int, nargs=2, default=(1, 4))
-    blocked_parser.add_argument("--order", type=int, nargs=2, default=(1, 0))
-    blocked_parser.add_argument("--blockShape", type=int, nargs=2, default=(128, 64),
-                        help='2D block size (dim0, dim1) to override the inferred size from blocked information')
+    # blocked layout parameters
+    blocked_parser = subparsers.add_parser(
+        "blocked",
+        help="plot blocked layout for global memory access",
+        formatter_class=lambda prog: OneLineFormatter(prog, max_help_position=40),
+    )
+    # tensor shapes
+    blocked_parser.add_argument("-r", "--rowName", type=str, default="M", metavar="ROW", help='tensor dim0 name')
+    blocked_parser.add_argument("-c", "--colName", type=str, default="K", metavar="COL", help='tensor dim1 name')
+    blocked_parser.add_argument("-B", "--matrixB", action="store_true", help='shortcut to plot operand B with dimension name of (K, N)')
+    blocked_parser.add_argument("-s", "--sizePerThread", type=int, nargs=2, default=(1, 4), metavar=("s0", "s1"),
+        help="how many elements each thread holds in the 2D block per CTA")
+    blocked_parser.add_argument("-t", "--threadsPerWarp", type=int, nargs=2, default=(16, 4), metavar=("t0", "t1"),
+        help="how thread is partitioned into a 2D grid in a warp with 64 threads")
+    blocked_parser.add_argument("-w", "--warpsPerCTA", type=int, nargs=2, default=(1, 4), metavar=("w0", "w1"),
+        help="how warps tile a CTA")
+    blocked_parser.add_argument("-o", "--order", type=int, nargs=2, default=(1, 0), metavar=("minor", "major"),
+        help="order from most minor to most major")
+    blocked_parser.add_argument("-b", "--blockShape", type=int, nargs=2, metavar=("b0", "b1"),
+        help='block size (dim0, dim1) of the tile. If not specified it presumably equals to the shape of CTA')
     ## dot layout parameters
-    dot_parser = subparsers.add_parser("dot", help="plot dot layout for MFMA")
-    dot_parser.add_argument("--warpsPerCTA", type=int, nargs=2, default=(1, 4))
-    dot_parser.add_argument("--dotShape", type=int, nargs=3, default=(32, 128, 64), help='Dot op shape in the form of M,N,K')
-    dot_parser.add_argument("--nonKDim", type=int, default=16, choices=[16, 32], help='mfma instruction dim')
+    dot_parser = subparsers.add_parser(
+        "dot",
+        help="plot dot layout for MFMA",
+        formatter_class=lambda prog: OneLineFormatter(prog, max_help_position=50),
+    )
+    dot_parser.add_argument("--dotShape", type=int, nargs=3, default=(32, 128, 64), metavar=("M", "N", "K"),
+                            help='Dot op shape in the form of M, N, K')
+    dot_parser.add_argument("--warpsPerCTA", type=int, nargs=2, default=(1, 4), metavar=("w0", "w1"),
+                            help="how warps tile the dot result matrix")
+    dot_parser.add_argument("--nonKDim", type=int, default=16, choices=[16, 32],
+                            help='mfma instruction dimension of M/N')
     dot_parser.add_argument("--kWidth", type=int, default=4, choices=[4, 8, 16, 32],
-                        help='number of contiguous elements per thread')
+                        help='number of contiguous elements each thread owns during MFMA')
     dot_parser.add_argument("--kGroup", type=int, default=1, choices=[1, 2],
                         help='total number of elements / kWidth per mfma instruction')
-    dot_parser.add_argument("--dtype-a", type=str, default='fp16',
+    dot_parser.add_argument("--dtypeA", type=str, default='fp16',
                         choices=['fp16', 'bf16', 'fp8', 'bf8', 'fp6', 'bf6', 'f4',
                                  'i8'], help='element type of operand A')
-    dot_parser.add_argument("--dtype-b", type=str, default='fp16',
+    dot_parser.add_argument("--dtypeB", type=str, default='fp16',
                         choices=['fp16', 'bf16', 'fp8', 'bf8', 'fp6', 'bf6', 'f4',
                                  'i8'], help='element type of operand B')
-    dot_parser.add_argument("--mfmaTrans", action='store_true', default=False, help='If set, then use mfma.trans layout')
+    dot_parser.add_argument("--mfmaTrans", action='store_true', default=False,
+                            help='If set, then use mfma.trans layout')
     dot_parser.add_argument("--scale", action='store_true', default=False,
                         help='If set, plot the scale tensor for mfma_f8f6f4 instructions')
     ## LDS access parameters
-    lds_parser = subparsers.add_parser("lds", help="plot LDS (shared memory) layout")
+    lds_parser = subparsers.add_parser(
+        "lds",
+        help="plot LDS (shared memory) layout",
+        formatter_class=lambda prog: OneLineFormatter(prog, max_help_position=50),
+        )
     lds_parser.add_argument("--tensorShape", type=int, nargs=2, default=(128, 64),
                         help='2D block shape in the form of (dim0, dim1)')
     lds_parser.add_argument("--kWidth", type=int, default=4, choices=[4, 8, 16, 32],
@@ -66,9 +86,9 @@ def parse_args():
                                  'i8'], help='element type of tensor to be stored in LDS')
     lds_parser.add_argument("--nonKDim", type=int, default=16, choices=[16, 32], help='mfma instruction dim')
     lds_parser.add_argument("--banks", type=int, default=32, choices=[32, 64], help='choose the number of banks in LDS')
-    lds_parser.add_argument("--lds-layout", type=str, default="none", choices=['swizzle', 'padding', 'none'],
+    lds_parser.add_argument("--layout", type=str, default="none", choices=['swizzle', 'padding', 'none'],
                         help='choose the LDS data layout')
-    lds_parser.add_argument("--lds-access", type=str, default="none", choices=['read', 'write', 'none'],
+    lds_parser.add_argument("--access", type=str, default="none", choices=['read', 'write', 'none'],
                         help='choose LDS access mode')
     lds_parser.add_argument("--mnContig", action='store_true', default=False,
                         help='If set, the tensor is K x N and n-contig')
@@ -82,15 +102,26 @@ def parse_args():
     wmma_parser = subparsers.add_parser("wmma", help="plot dot layout for wmma")
     wmma_parser.add_argument("--wave-size", type=int, default=32, choices=[32, 64], help='choose the wmma instruction mode')
 
-    args = parser.parse_args()
+    # workaround for top-level parser's flag passes in after subparser
+    args, remaining = top_parser.parse_known_args()
+    args = top_parser.parse_args(remaining, namespace=args)
 
     return args
 
 
 def main():
     args = parse_args()
+    force = args.force
     ofilename = args.output
     keepSrc = args.keep
+
+    if os.path.exists(f"{ofilename}.pdf"):
+        if not force:
+            print(f"{ofilename}.pdf exists already. Please use --force to overwrite or change output name")
+            sys.exit(0)
+        else:
+            print(f"{ofilename}.pdf exists but overwritten!")
+
 
     match args.plot_type:
         case "blocked":
@@ -104,15 +135,18 @@ def main():
         case _:
             raise NotImplementedError(f"Only blocked, dot, lds and wmma supported, you entered {args.plot_type}")
 
-    run_bash_command(f"pdflatex -jobname {ofilename} myplot.tex")
-    print(f"plot saved in {ofilename}.pdf")
-
-    # Remove auxiliary files
-    os.remove(f"{ofilename}.aux")
-    os.remove(f"{ofilename}.log")
-    if not keepSrc:
-        os.remove("myplot.tex")
-        run_bash_command("rm -rf ./auto")
+    ret = run_bash_command(f"pdflatex -halt-on-error -jobname {ofilename} myplot.tex")
+    if ret == 0:
+        print(f"plot saved in {ofilename}.pdf")
+        # Remove auxiliary files
+        os.remove(f"{ofilename}.aux")
+        os.remove(f"{ofilename}.log")
+        if not keepSrc:
+            os.remove("myplot.tex")
+            run_bash_command("rm -rf ./auto")
+    else:
+        print("pdflatex has problem generating the pdf file. Check myplot.log for the error!!!")
+        sys.exit(1)
 
 
 if __name__ == '__main__':
